Rewrite DotnetBuildConfiguration with optional fields
This removes the explicit default VerbosityLevel configuration (which in turn will now rely on the .NET SDK default).

While the default build configuration value ("Release") is moved out of the DotnetBuildpackConfiguration struct to make the data model more consistent. The same value is still set explicitly in the main function. We currently need this value to compute the expected output path (for process type detection).