Improve test assertion
When the expected value is parsed using VersionReq, tests may pass when the expected version requirement string doesn't match the parsed value. Updating the test to ensure that we compare the actual (parsed) version requirement's string representation with the expected value.

This improves the assertion for the (default) behavior, when no rollForward policy is defined (and the input text is passed as-is), and makes it clear this is parsed as a "caret" comparison.