Rewrite NuGet cache restore evaluation
NugetCacheLayerMetadata#restore_count is intended to track the number of times the layer has been restored, so it makes more sense to set this value to 0 when the layer has just been created.

This should improve readability by reducing the ambiguity between "number of NuGet package restores" (which are part of every build/publish), and "cache layer restores". Updated the log message wording accordingly.