Drop default build configuration output
When build configuration is specified it'll be evident from the command line, which is already printed to the buildpack output.

When unspecified, the build configuration value is excluded from the executed command, and the user can safely assume it's the default value for `dotnet publish` and consistent with the local behavior.