Use only solution file name for test command
This changes the test launch process command to use the solution file name rather than the absolute path.

The motivation for this change is the same as https://github.com/heroku/buildpacks-dotnet/pull/138.

In this case, however, there's no need to also change directory, as we're only testing the solution file, which is required to be in the working directory.