Trim assembly name project property value
The .NET CLI will simply ignore an AssemblyName property containing only whitespace, so we trim the value before checking if it's empty.

The value isn't trimmed before setting the assembly name value on the metadata, so "  foo " will be valid in this context. This also mirrors the .NET CLI behavior, which will later fail to publish projects with starting/ trailing whitespace in the assembly name/