Amend existing changelog entry
The changes introduced in this PR could be considered a fix, or a change to the existing logic. However, this change has no impact on the behavior users of the buildpack currently experience, as #310 hasn't been released (and there were no way to specify a nested solution file prior to that).

Amending the existing, unreleased changelog entry keeps the changelog focused and avoids potential confusion (e.g. a user might read it and think it affects already released behavior).